requirements.txt
selenium>=4.0.0
openpyxl>=3.0.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
